Resolved PR Comments
